+++ source/extensions/omni.isaac.sensor/python/tests/__init__.py
@@ -12,6 +12,7 @@ import sys
 from .test_camera_sensor import *
 from .test_contact_sensor import *
 from .test_contact_sensor_wrapper import *
+from .test_effort_sensor import *
 from .test_imu_sensor import *
 from .test_imu_sensor_wrapper import *
 from .test_lidar_rtx import *